Use `deactivate` to tell if in virtual environment
This patch makes it so the combined definition of $VIRTUAL_ENV and
existence of the shell function deactivate environment are used to tell
whether a virtual environment is active.

This was done to fix some issues where the plugin collided with other
virtual environment managers like Direnv. Direnv sets $VIRTUAL_ENV but
crucially does not create the disable shell function, which makes sense,
as Direnv manages that for the user.

Previosly, the plugin would not discover the virtual environment created
by Direnv and (since Direnv set VIRTUAL_ENV) decide that it should leave
the virtual environment. This would print an error message as deactivate
is not defined by Direnv.